[TASK] Restructure functional tests
Tests from "Cases" folder are moved to the appropriate location
in Core or ViewHelpers. In some cases, test files are combined into
one.

phpunit result before change:
Tests: 1288, Assertions: 2709, Skipped: 1.

phpunit result after change:
Tests: 1288, Assertions: 2709, Skipped: 1.
