+++ tests/Functional/Core/Parser/WhitespaceToleranceTest.php
