+++ tests/Functional/Core/Rendering/RecursiveRenderingTest.php
@@ -7,7 +7,7 @@ declare(strict_types=1);
  * See LICENSE.txt that was shipped with this package.
  */
 
-namespace TYPO3Fluid\Fluid\Tests\Functional\Cases\Rendering;
+namespace TYPO3Fluid\Fluid\Tests\Functional\Core\Rendering;
 
 use PHPUnit\Framework\Attributes\Test;
 use TYPO3Fluid\Fluid\Tests\Functional\AbstractFunctionalTestCase;
