+++ tests/Functional/Core/ViewHelper/ConditionViewHelperTest.php
@@ -7,7 +7,7 @@ declare(strict_types=1);
  * See LICENSE.txt that was shipped with this package.
  */
 
-namespace TYPO3Fluid\Fluid\Tests\Functional\Cases\Conditions;
+namespace TYPO3Fluid\Fluid\Tests\Functional\Core\ViewHelper;
 
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\Test;
@@ -15,8 +15,56 @@ use TYPO3Fluid\Fluid\Tests\Functional\AbstractFunctionalTestCase;
 use TYPO3Fluid\Fluid\Tests\Functional\Fixtures\Various\UserWithToString;
 use TYPO3Fluid\Fluid\View\TemplateView;
 
-final class VariableConditionsTest extends AbstractFunctionalTestCase
+final class ConditionViewHelperTest extends AbstractFunctionalTestCase
 {
+    public static function basicConditionDataProvider(): array
+    {
+        return [
+            ['1 == 1', true],
+            ['1 != 2', true],
+            ['1 == 2', false],
+            ['1 === 1', true],
+            ['\'foo\' == 0', false],
+            ['1.1 >= \'foo\'', false],
+            ['\'String containing word \"false\" in text\'', true],
+            ['\'  FALSE  \'', true],
+            ['\'foo\' > 0', true],
+            ['FALSE', false],
+            ['(FALSE || (FALSE || 1)', true],
+            ['(FALSE or (FALSE or 1)', true],
+
+            // integers
+            ['13 == \'13\'', true],
+            ['13 === \'13\'', false],
+
+            // floats
+            ['13.37 == \'13.37\'', true],
+            ['13.37 === \'13.37\'', false],
+
+            // groups
+            ['(1 && (\'foo\' == \'foo\') && (TRUE || 1)) && 0 != 1', true],
+            ['(1 && (\'foo\' == \'foo\') && (TRUE || 1)) && 0 != 1 && FALSE', false],
+        ];
+    }
+
+    #[DataProvider('basicConditionDataProvider')]
+    #[Test]
+    public function basicCondition(string $source, bool $expected): void
+    {
+        $source = '<f:if condition="' . $source . '" then="yes" else="no" />';
+        $expected = $expected === true ? 'yes' : 'no';
+
+        $view = new TemplateView();
+        $view->getRenderingContext()->setCache(self::$cache);
+        $view->getRenderingContext()->getTemplatePaths()->setTemplateSource($source);
+        self::assertSame($expected, $view->render());
+
+        $view = new TemplateView();
+        $view->getRenderingContext()->setCache(self::$cache);
+        $view->getRenderingContext()->getTemplatePaths()->setTemplateSource($source);
+        self::assertSame($expected, $view->render());
+    }
+
     public static function variableConditionDataProvider(): array
     {
         $user1 = new UserWithToString('foobar');
@@ -93,7 +141,7 @@ final class VariableConditionsTest extends AbstractFunctionalTestCase
 
     #[DataProvider('variableConditionDataProvider')]
     #[Test]
-    public function basicCondition(string $source, bool $expected, array $variables): void
+    public function variableCondition(string $source, bool $expected, array $variables): void
     {
         $source = '<f:if condition="' . $source . '" then="yes" else="no" />';
         $expected = $expected === true ? 'yes' : 'no';
