+++ tests/Functional/Core/ViewHelper/ViewHelperEscapingTest.php
@@ -7,7 +7,7 @@ declare(strict_types=1);
  * See LICENSE.txt that was shipped with this package.
  */
 
-namespace TYPO3Fluid\Fluid\Tests\Functional\Cases\Escaping;
+namespace TYPO3Fluid\Fluid\Tests\Functional\Core\ViewHelper;
 
 use TYPO3Fluid\Fluid\Core\Parser\Configuration;
 use TYPO3Fluid\Fluid\Core\Parser\Interceptor\Escape;
