+++ tests/Functional/Fixtures/Various/Objects/WithProperties.php
@@ -7,7 +7,7 @@ declare(strict_types=1);
  * See LICENSE.txt that was shipped with this package.
  */
 
-namespace TYPO3Fluid\Fluid\Tests\Functional\Cases\Rendering\Fixtures\Objects;
+namespace TYPO3Fluid\Fluid\Tests\Functional\Fixtures\Various\Objects;
 
 class WithProperties
 {
